+++ WordPress-Extra/ruleset.xml
@@ -87,7 +87,7 @@
 
 	<rule ref="WordPress.PHP.DevelopmentFunctions"/>
 	<rule ref="WordPress.PHP.DiscouragedPHPFunctions"/>
-	<rule ref="WordPress.PHP.PerformanceEnhancementPHPFunctions"/>
+	<rule ref="WordPress.CodeAnalysis.PerformanceEnhancementPHPFunctions"/>
 	<rule ref="WordPress.WP.DeprecatedFunctions"/>
 	<rule ref="WordPress.WP.DeprecatedClasses"/>
 	<rule ref="WordPress.WP.DeprecatedParameters"/>
