+++ WordPress/Tests/CodeAnalysis/PerformantPHPFunctionsUnitTest.inc
